Fixed: Bruno didnt work because of the data corrupted by the merges of the branches to the Main

+++ donations-system-app/Dtos/RequestDto.cs
@@ -5,5 +5,5 @@ public class RequestDto
     public string InstitutionName { get; set; }
     public string Description { get; set; }
     public string NeedType { get; set; }
-    public decimal RequestedQuantity { get; set; }
+    public double RequestedQuantity { get; set; }
 }
